+++ src/Authorization/src/Roles/RoleNotFoundException.php
@@ -1,0 +1,23 @@
+<?php
+
+/**
+ * Aphiria
+ *
+ * @link      https://www.aphiria.com
+ * @copyright Copyright (C) 2019 David Young
+ * @license   https://github.com/aphiria/Opulence/blob/master/LICENSE.md
+ */
+
+declare(strict_types=1);
+
+namespace Opulence\Authorization\Roles;
+
+use Exception;
+
+/**
+ * Defines the exception that's thrown when a role is not found
+ */
+final class RoleNotFoundException extends Exception
+{
+    // Do not do anything
+}