+++ src/QueryBuilders/src/SelectQuery.php
@@ -381,7 +381,7 @@ class SelectQuery extends Query
             } elseif (is_string($condition)) {
                 $conditionExpressions[] = $condition;
             } else {
-                throw new InvalidArgumentException('Condition must either be string or ICondition object');
+                throw new InvalidArgumentException('Condition must either be string or ' . ICondition::class . ' object');
             }
         }