+++ src/QueryBuilders/src/UpdateQuery.php
@@ -152,7 +152,7 @@ class UpdateQuery extends Query
             } elseif (is_string($condition)) {
                 $conditionExpressions[] = $condition;
             } else {
-                throw new InvalidArgumentException('Condition must either be string or ICondition object');
+                throw new InvalidArgumentException('Condition must either be string or ' . ICondition::class . ' object');
             }
         }